pyproject.toml
[tool.poetry]
name = "flask-rest-api-template"
version = "0.1.0"
description = "Template to create REST APIs with flask"
authors = ["Henke Adolfsson <catears13@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
flask_restful = "^0.3.8"
gunicorn = "^20.0.4"
dynaconf = "^3.0.0"
pymongo = "^3.11.0"
python-redis = "^0.2.2"
rq = "^1.5.0"
loguru = "^0.5.1"
jsonschema = "^3.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
doit = "^0.32.0"
pytest-cov = "^2.10.0"

[tool.pytest.ini_options]
testpaths = [
  "tests",
]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

